requirements.txt
requests==2.31.0
pandas==2.1.1
openpyxl==3.1.2
Pillow==10.1.0
google-generativeai>=0.4.0 